# _oas/openapi.yaml
openapi: 3.1.0
info:
  title: vega api
  description: vega api
  version: 1.0.0
servers:
  - url: 'http://vega.localhost'
paths:
  /health:
    get:
      operationId: "getHealth"
      description: "get health"
      responses:
        200:
          description: Health
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/Health"
        default:
          $ref:  "#/components/responses/Error"
  /applications:
    get:
      operationId: "getApplications"
      description: "get application list"
      responses:
        200:
          description: Application list
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ApplicationList"
        default:
          $ref:  "#/components/responses/Error"
  /applications/{name}:
    get:
      operationId: "getApplication"
      description: "get application"
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: "Application name"
      responses:
        200:
          description: Application
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/ApplicationSummary"
        default:
          $ref:  "#/components/responses/Error"
components:
  schemas:
    # Error-related schemas.
    TraceID:
      type: string
      description: W3C trace-id
      pattern: "[[:xdigit:]]{32}"
      example: 0af7651916cd43dd8448eb211c80319c
      externalDocs:
        url: "https://www.w3.org/TR/trace-context/#trace-id"
        description: "W3C Trace Context specification"
    SpanID:
      type: string
      description: W3C parent-id (span)
      pattern: "[[:xdigit:]]{16}"
      example: b7ad6b7169203331
      externalDocs:
        url: "https://www.w3.org/TR/trace-context/#parent-id"
        description: "W3C Trace Context specification"
    Error:
      title: Structured error
      description: Error occurred while processing request
      externalDocs:
        url: "https://pfm.pages.gitlab.corp.mail.ru/docs/dev/spec/v2/errors"
        description: "Structured error specification"
      type: object
      required:
        - error_message
        - body
      properties:
        error_message:
          type: string
          description: "Human-readable error message"
          example: "Something went wrong"
        trace_id:
          $ref: "#/components/schemas/TraceID"
        span_id:
          $ref: "#/components/schemas/SpanID"

    # Health-related schemas.
    Health:
      type: object
      required:
        - status
        - version
        - commit
        - build_date
      properties:
        status:
          type: string
          description: "Health status"
          example: "ok"
        version:
          type: string
          description: "Service version"
          example: "1.0.0"
        commit:
          type: string
          description: "Service commit"
          example: "c1b2d3f4"
        build_date:
          type: string
          description: "Service build date"
          example: "2022-01-01T00:00:00Z"
          format: date-time

    # API.
    Application:
      type: object
      required:
        - name
        - namespace
      properties:
        name:
          type: string
          description: "Application name"
          example: "api"
        namespace:
          type: string
          description: "Application namespace"
          example: "vega"

    Pod:
      type: object
      required:
        - name
        - namespace
        - status
      properties:
        name:
          type: string
          description: "Pod name"
          example: "api-123456"
        namespace:
          type: string
          description: "Pod namespace"
          example: "vega"
        status:
          type: string
          description: "Pod status"
          example: "Running"

    ApplicationSummary:
      type: object
      required:
        - name
        - namespace
      properties:
        name:
          type: string
          description: "Application name"
          example: "api"
        namespace:
          type: string
          description: "Application namespace"
          example: "vega"
        pods:
          type: array
          items:
            $ref: "#/components/schemas/Pod"


    ApplicationList:
      type: array
      items:
        $ref: "#/components/schemas/Application"

  responses:
    Error:
      description: Structured error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"